OrderItem : refactored ID datatype to long

+++ src/main/java/com/kuna_backend/repositories/OrderItemsRepository.java
@@ -3,5 +3,5 @@ package com.kuna_backend.repositories;
 import com.kuna_backend.models.OrderItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderItemsRepository extends JpaRepository<OrderItem, Integer> {
+public interface OrderItemsRepository extends JpaRepository<OrderItem, Long> {
 }
